+++ app/helpers/dashboard_helper.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+# Dashboard helper class
+module DashboardHelper
+end